create a runtime panel to track runtimes
