+++ crates/repl/Cargo.toml
@@ -17,6 +17,7 @@ anyhow.workspace = true
 alacritty_terminal.workspace = true
 async-dispatcher.workspace = true
 base64.workspace = true
+client.workspace = true
 collections.workspace = true
 editor.workspace = true
 gpui.workspace = true
