+++ crates/zed/src/main.rs
@@ -221,7 +221,7 @@ fn init_ui(app_state: Arc<AppState>, cx: &mut AppContext) -> Result<()> {
 
     assistant::init(app_state.fs.clone(), app_state.client.clone(), cx);
 
-    repl::init(app_state.fs.clone(), cx);
+    repl::init(cx);
 
     cx.observe_global::<SettingsStore>({
         let languages = app_state.languages.clone();
